Use curly quotes throughout
The original text used curly single and double quotes, but the
conversion process converted all of them to straight quotes.

This commit converts the single and double straight quote
characters *back* to curly single and double quotes where they
should be converted. This did *not* convert anything in `~~~~`
markdown ranges.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>