Add section on integer overflow - fixes #16
We briefly discuss integer overflow for *signed* integer overflow
in C and C++, but the problem of integer overflow and underflow
is much broader.

This commit adds a new section about integer overflow & underflow
in general. It includes a discussion about different programming
languages, as well as a story time example.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>