Add links to some OpenSSF projects, fixes #19
This course was one of the first things released by the OpenSSF,
so of course it didn't mention many other things done by the OpenSSF.

This commit add links to several ready-to-use OpenSSF projects.
It does not add *all* OpenSSF projects. For example, I didn't Alpha-Omega;
a typical developer doesn't need to consult Alpha-Omega for anything.
I also didn't add SLSA; SLSA is self-described as version 0.1, so
while SLSA has great promise, I didn't think it was yet ready to add
to this particular course. I look forward to when it's ready.

I didn't add sigstore; it obviously can be used, but it's still in
an early stage as well, so I didn't add it. That's probably the closest
call that could be argued either way.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>