Note 2021 CWE Top 25 ranks 1-5
We want to add references to the 2021 CWE top 25
instead of just the 2019 version.

This adds citations to the 2021 CWE top 25 for its
ranks 1-5. We already cover them (as we do in most cases),
so we just need to note them here.
If we can do them all in time we can update the table
near the end, but no matter what we can update things here.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>