Yes, Windows has lots of link vulnerabilities
I did a quick search at CVEs by CWE vulnerabilities,
searching for CWE-59:
https://www.cvedetails.com/vulnerability-list/cweid-59/vulnerabilities.html

What I found was that link vulnerabilities on Windows happen
quite often. I had expected them to be rare because creating links
on Windows is a privileged operation. That turns out to be untrue!
So make it clear this really is a problem for Windows as well.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>