Update link handling (including Story Time)
Update the link handling text. This includes a
"Story Time" entry. This tutorial text was really hard to
write because it seemed so abstract; the story time is
really needed here because the worked example makes it more
complete. We worked to simplify the otherwise-complex example.
There are a lot of CVEs in this category, but many of them
are really hard to explain or so specialized that a developer
might just ignore the issue.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>