Clarify sticky bit
The exact semantics of the sticky bit vary slightly between
Unix-like systems. See:
https://en.wikipedia.org/wiki/Sticky_bit

The goal here isn't to enumerate the variation, but instead
recommend that people *use* the sticky bit, and that the
semantics described here are the Linux semantics.

A few other tweaks about sticky bits were also made.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>