Discuss *parameterized* vs. *prepared* statements
The term "prepared statements" and "parameterized statements"
are treated as synonyms in many APIs. However, that is *not*
always true - there are APIs where they are *different*.
In those cases prepared statements generally *include*
parameterized statement capabilities, but not the other way around.

Switching to this terminology also makes it easier to (briefly)
discuss what to do with stored procedures.

I made a lot of changes in the SQL injection section, including a
number of additions, but SQL injection vulnerabilities are so
important that I think this is worth doing. I tried to give
specifics without being overwhelming. I added a few other
language-specific examples, again because SQL injection is such
a common problem that I want to increase the likelihood of
people seeing how to use them in a language they know.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>