Italicize CWE & OWASP list titles. Fixes #73
Italicize 20.. CWE Top 25 and 20.. OWASP Top 10 titles.
Note that the CWE Top 25 title sometimes includes
"Most Dangerous Software Errors". In one case it was already
italicized, so we revert the doubled asterisks, and
we also fix a missing space after an "and" before italicizing it.

This is the result of:

~~~~sh
cat secure_software_development_fundamentals.md | \
  sed -E 's/(20.. CWE Top 25( Most Dangerous Software Errors)?)/\*\1\*/g' | \
  sed -E \
    's/\*(\*20.. CWE Top 25( Most Dangerous Software Errors)?\*)\*/\1/g' | \
  sed -E 's/(20.. OWASP Top 10)/\*\1\*/g' | \
  sed -E 's/ and\*/ and \*/g' > ,1
  mv ,1 secure_software_development_fundamentals.md
~~~~

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>