Update evaluating OSS list
The OpenSSF Best Practices WG has approved its
[Concise Guide for Evaluating Open Source Software 2022-08-26](https://github.com/ossf/wg-best-practices-os-developers/blob/main/docs/Concise-Guide-for-Evaluating-Open-Source-Software.md#readme).

So let's use it instead.

This isn't as big a change as it looks. The first draft of the
concise guide was mainly the content from this course.
However, many people have made various improvements, so let's
incorporate this improved list back in the course.

The original course list was for any reused software, while this
new list focuses on OSS. That's okay; most reused software is OSS.
however, that means we need to tweak the text to match.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>