Clean up application-side text
This example includes some backquotes, which are both
part of the text but are also used in markdown to indicate
teletype font (tt). This led to problems. Fix that, and while
we're at it, clarify the text. This is challenging to explain,
so adding a few asides should make it easier to understand.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>