Make all section names unique and enforce it
The HTML anchor names for sections are automatically generated
from their section heading text. That's fine, but if we want to make
sure we can refer to any section, we need to ensure that every
section heading's text is unique.

This modifies the context to make every section heading unique
(after changing the quiz headings we were close), and modifies
our markdownlint configuration so we'll reject future changes
that create duplicate section text entries. It's still possible
for multiple locations to end up with the same anchor name
(e.g., if they differ only in characters that aren't used to create
the anchor name), but in practice this is highly unlikely and
we can address this by reviewing changes before accepting them.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>