Add Ashley Madison Story Time
Add Ashley Madison Story Time about using MD5 for passwords.

The Ashley Madison data breach
story is more complicated, e.g., the mere fact that someone
was in the database was a problem, and people could create
accounts for others' email address without verification:
https://en.wikipedia.org/wiki/Ashley_Madison_data_breach

However, if we focus on just the use of MD5, that part of
the story is much easier to explain, and it's a helpful example
that got a lot of media attention. It also helps clarify that
"storing in both secure AND insecure formats" is a bad idea.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>