First draft text of proposed video clips
This creates simple text to introduce major sections with
short video clips. I intend for one to be done by CRob.

One feedback item we got was that many people wanted a little
more media (video) to break the monotony. The problem is that
creating videos takes a lot of time, and more importantly, makes
editing the material hard. We might be able to resolve
that with AI, but that's a separate discussion. The proposed
solution is to have short video clips that introduce
chapters - they are unlikely to need updating, yet they'll help
break up the material.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>