Update the video intro text to eliminate duplication
LF Education & Training doesn't want the video text to *also*
be displayed as duplicate text, so let's mark it as "video"
but not duplicate it. I also tweaked the text to try to briefly
explain *why* certain chapters were important (so hopefully
the learner will be motivated to learn it).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>