Expand the Tux ECB image example
While cryptographers know the image and can imagine the original and
how it looks in ECB, many who start their path cannot imagine that process.
This commit updates the images to show the original (from wikipedia - converted
to png), the ECB-encrypted, and one encrypted with CTR (mentioned as GCM to
simplify the story, as CTR is not yet covered).

Signed-off-by: Nikos Mavrogiannopoulos <nikos.mavrogiannopoulos@assaabloy.com>