Add more labs, move them after quizzes
This adds more of the labs.

Also, I realized that the labs should be *after* the
quizzes. The quizzes focus on *recognizing* correct answers,
which is easier than *creating* your own. So the more logical
order in general is to do a quiz first, then note lab(s),
in cases where both are present.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>