Fix constant-time section (wrong Java method name)
This fixes an error reported here:
https://forum.linuxfoundation.org/discussion/866738/constant-time-algorithms-for-java
by Simon (sipaynehsl).

Since I'm already there, I added C#/.NET and some links
(adding the links is helpful, *AND* it helps everyone verify that
the given answers are correct). I changed the module and method
names here to Courier font; we should probably do that everywhere
eventually, but so this is a start.

Thank you!

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>